+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 ## Changelog
 
+### Version 1.8.3
+
+- Fix for splashscreen/icon viewing on Windows
+- Settings for live reload, external IP and preview in editor now displayed
+- Hide progress dialogs for running on iOS and Android
+- Fix for mono repos when ionic cli is installed and capacitor sync is called
+- Fix for mono repos for projects neither Capacitor or Cordova
+- When serving for web, if the External IP address is turned on it will use it
+- Animated icons for running tasks
+
+### Version 1.7.9
+
+- Login and Signup are now fixed
+- Cordova based Ionic Enterprise plugins are no longer flagged for review
+
+### Version 1.7.5
+
+- Live reload now works with Auth Connect if version is at least v3.9.4
+
 ### Version 1.7.4
 
 - Fix to allow Capacitor integrations to web projects by running ionic init